fix: remove icon projects;
