+++ frontend/app/admin/(admin)/news/page.tsx
