+++ frontend/app/admin/(admin)/people/page.tsx
