+++ frontend/app/admin/(admin)/settings/page.tsx
