+++ frontend/app/admin/layout.tsx
@@ -1,21 +1,3 @@
-import React from 'react';
-import { Metadata } from 'next';
-import { AdminLayout } from '@/vievs/admin-pages/AdminLayout';
-
-export const metadata: Metadata = {
-  title: 'Адмін панель',
-  description: 'Адмін панель сайту',
-};
-
-export default function Layout(
-  {children}
-  : {
-    children: React.ReactNode;
-  }): React.JSX.Element
- {
-  return (
-    <AdminLayout>
-      {children}
-    </AdminLayout>
-  );
-}
+export default function layout( {children} : {
+  children: React.ReactNode;
+}): JSX.Element { return <>{children}</> };
