+++ frontend/app/layout.tsx
@@ -8,6 +8,9 @@ import { Footer } from '@/widgets/Footer';
 export const metadata: Metadata = {
   title: 'My App',
   description: 'Description of my app',
+  icons: {
+    icon: ['/LogoDark.svg'],
+  },
 };
 
 export default function RootLayout({
@@ -16,7 +19,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
+    <html lang="uk-UA">
       <body className="bg-base-text_accent">
         <Header />
         <main>{children}</main>
