+++ frontend/app/thank/page.tsx
@@ -4,9 +4,6 @@ import { ThankPage } from '@/vievs/ThankPage';
 export const metadata: Metadata = {
   title: 'Дякуємо Вам!',
   description: 'Інформаційна сторінка подяки',
-  icons: {
-    icon: ['/LogoDark.svg'],
-  },
 };
 
 export default function thankPage() {
