+++ frontend/src/vievs/admin-pages/AdminLayout/AdminLayout.tsx
@@ -6,24 +6,10 @@ export function AdminLayout({children}
   : {
     children: React.ReactNode;
   }): React.JSX.Element
-  {
+  { 
   return (
-    <div className='flex flex-row gap-8 p-10 '>
-      <div className='w-[236px] bg-base-text_dark rounded-[30px] p-6'>
-        <Dashboard/>
-        {/* <ul className='w-full flex flex-col gap-4 items-center'>
-          <li>
-            Адмін панель
-          </li>
-          <li>
-            Адмін панель
-          </li>
-          <li>
-            Адмін панель
-          </li>
-        </ul> */}
-      </div>
-
+    <div className='flex flex-row gap-8 p-5 lg:p-10'>
+      <Dashboard/>
       <div className='mim-h-full w-full'>
        {children}
       </div>
