+++ frontend/src/vievs/admin-pages/UnderDevelopment/ui/UnderDevelopment.tsx
@@ -5,7 +5,7 @@ import { Text } from "@/shared/ui/Text";
 
 export function UnderDevelopment(): JSX.Element {
   return (
-    <div className="flex h-full w-full flex-col items-center justify-center gap-4 p-4">
+    <div className="h-full w-full flex flex-col items-center justify-center gap-4 p-4">
     <Text
       Tag="h1"
       textType="Desktop/H2"
@@ -21,7 +21,6 @@ export function UnderDevelopment(): JSX.Element {
         textType="Desktop/H3"
         text={'Ця сторінка в розробці'}
         font="sans"
-        //align="center"
         color="base/text"
         className="font-normal text-[28px]"
       />
@@ -31,6 +30,7 @@ export function UnderDevelopment(): JSX.Element {
       variant="arrowTextBlue"
       text={'На головну'}
       to="/"
+      className='duration-300 hover:opacity-70'
     />
   </div>
   )
